requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
pillow-heif==0.15.0
numpy==1.26.3
python-dotenv==1.0.0
aiofiles==23.2.1

# Background Removal
rembg==2.0.56
carvekit==4.1.2
backgroundremover==0.2.7

# Deep Learning
torch==2.1.2
torchvision==0.16.2

# Image Processing & Enhancement
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
scikit-image==0.22.0

# Vector Graphics
pycairo==1.26.0
cairosvg==2.7.1
svgwrite==1.4.3

# Additional utilities
scipy==1.11.4
